Refactor some stuff, fix fonts, assert vao/vbo

+++ src/rico_font.h
@@ -35,8 +35,7 @@ struct rico_font
 };
 
 global const char *font_name(struct rico_font *font);
-void font_render(u32 *mesh_id, u32 *tex_id, struct rico_font *font, int x,
-                 int y, struct vec4 bg, const char *text,
+void font_render(u32 *mesh_id, u32 *tex_id, struct rico_font *font, float x,
+                 float y, struct vec4 bg, const char *text,
                  const char *mesh_name);
-
 #endif
